Improve extended DeviceResource propagation
The first version of the method to find matching release actions for one acquire chose the next release with the same resource. This was not correct in the case of multiple acquires for the same resource before releasing it (with multiple releases). The IntBIIS model actually supports such concurrent usage, so the algorithm is improved and counts the resource usage now; only if the number of releases and acquires matches, all actions in between are marked (and those actions between could also be acquires and releases for the same resource).