Extend propagation algorithm with ActorStep -> ActorStep step
If an ActorStep is marked for modification, all directly following ActorSteps are also marked. The semantic is quite weak in that case, as we don't have information what is changed about the ActorStep, so its successors are chosen as an approximation. If you know exactly that a DataObject changes (which also affects ActorSteps), you should mark it instead and all ActorSteps with that DataObject will be marked.