Fix check for added data dependencies
The algorithm creates a BPChangePropagationDueToDataDependencies element where different types of modifications can be added. If none is added, the element representing that propagation step should be removed instead of remaining without child in the tree. The method for that empty-check was inherited from the abstract propagation algorithm class (superclass for concrete algorithms of KAMP and KAMP+IntBIIS), but was based on KAMP's ChangePropagationDueToDataDependencies model element, not considering new types of modifications (new child elements) that could also be there in KAMP+IntBIIS. This commit fixes that problem by replacing the hard-coded check for child elements with  a generic EMF solution.