Adapt workplan derivation for ActorStep ->ActorStep and remove duplicates
The recent meta-model extension of ModifyActorSteps (it can have a ModifyActorStep reference now, representing the previous step in the ActorStep -> ActorStep propagation) is considered when determining the subactivities for the workplan. Furthermore, the performance should be improved and the number of duplicates reduced, as the former generic algorithm to find all modificationmarks of a certain type (involving a full model traversal for each) is replaced by hard-coded lookups of relevant modificationmarks in the repository. This has the disadvantage that in case of new kinds of propagation steps or new attribute (references) for InterBusinessProcessPropagation or BpChangePropagationDueToDataDependencies, slight adaptations have to be made to BPInternalModificationDerivation so that the new modificationmarks are included. As the major advantage, it is guaranteed that all modificationmarks from the tree are only considered once (especially relevant for child element modifications which also appear as top-level modifications at a different location).