Improve OperationSignature -> OperationSignature step
An older commit added the support for OperationToOperationDependencies when calculating the intra-component propagation. However, the old solution was non-deterministic, as it still used KAMP's visitedProvidedRoles and visitedRequiredRoles which guaranteed that all roles are only visited once during inter- and intra-component-propagation. Therefore, the order of visited roles influenced the marked signatures based on OperationToOperationDependencies. 
This commit adapts inter- and intra-component propagation so that roles can be visited multiple times and the OperationToOperationDependencies are fully utilized. Termination is still guaranteed by two mechanisms: Counting the number of visited AssemblyConnectors in the inter-component step and only adding ProvidedRoles that were not visited or whose modification mark had been changed in the intra-component step. Furthermore, the routines for that steps are refactored, especially to shorten the methods (extract sub-routines).