Extend example for improved ActorStep -> ActorStep propagation
To demonstrate the propagation between indirectly succeeding ActorSteps with data dependencies, some new DataObjects are introduced and the usagemodel is enhanced with additional steps. For the trickier parts of the propagation, a Loop and two Branches are added. Further parts of the file have been modified to make the usagemodel a valid one (adding missing required attributes etc.).