Extend propagation algorithm with Decision -> Requirement step
Similar to the propagation of modified Requirements to Decisions, this step goes the opposite direction, meaning modified Decision -> Requirement. The logic for checking if requirements and decisions are related is quite simple and nearly the same in this step as in the opposite direction (checking the triggeredBy EReference of the Decision if it contains the Requirement).