Enhance InterOption-/-Decision-/-RequirementPropagation transitively
The old lookups for options depending on options, requirements depending on requirements and decisions depending on decisions only considered direct successors. This commit introduces a transitive lookup which is even able to handle all three lookup types in one method by adding a parameter for the class of the result.