Extend propagation algorithm with Interface -> Signature step
The algorithm already contained a Signature -> Interface step. This commit adds the opposite direction, meaning that a user can mark an Interface when all Signatures of it should be marked. The KAMP+IntBIIS modificationmarks model is adapted accordingly (new kind of propagation step, because it does not fit in the existing categories [data dependencies, business process, inter-component, intra-component]).