Generalize modificationsmarks and componentInternalDepencies model
Before this commit, the models for the KAMP algorithm were related to the Operation... elements (-Signature, -Interface, -ProvidedRole, -RequiredRole, AssemblyConnector). This commit generalizes the modificationsmarks and componentInternalDepencies model, so they can also refer to any sub-type of Signature, Interface, ProvidedRole, RequiredRole, Connector. This involves some refactoring of the model and the algorithm as well, as the types of the ModifyXyz attributes are generalized. The drawback is a loss of type-safety (one additional instanceof in the KAMP/KAMP+IntBIIS algorithms, many instanceofs in LabelCustomizing). The major advantage is the extension of modifications to new types of architecture elements (especially EventType, EventGroup, SourceRole, SinkRole) without adding new elements to the modificationmarks model.