Extend propagation algorithm with EventGroup -> SourceRole/SinkRole/EventType step
Similar to the (Operation)Interface -> (Operation)RequiredRole/(Operation)ProvidedRole/(Operation)Signature propagation, this commit generalizes these steps to EventTypes (which are Signatures), EventGroups (which are Interfaces), SourceRoles (which are RequiredRoles) and SinkRoles (which are ProvidedRoles). This involves some type changes/generalizations in the KAMP algorithm and the ArchitectureModelLookup, but no new propagation steps.